Rename migration lib and prevent rendering while the migration is running
