+++ src/Expensify.js
@@ -12,7 +12,7 @@ import * as ActiveClientManager from './libs/ActiveClientManager';
 import ONYXKEYS from './ONYXKEYS';
 import styles from './styles/styles';
 import Log from './libs/Log';
-import MigrateOnyx from './libs/MigrateOnyx';
+import migrateOnyx from './libs/migrateOnyx';
 import {
     Route,
     Router,
@@ -73,13 +73,15 @@ class Expensify extends PureComponent {
         this.state = {
             isLoading: true,
             authToken: null,
+            isOnyxMigrated: false,
         };
     }
 
     componentDidMount() {
         // Run any Onyx schema migrations and then connect to Onyx
-        MigrateOnyx()
+        migrateOnyx()
             .then(() => {
+                this.setState({isOnyxMigrated: true});
                 Onyx.connect({
                     key: ONYXKEYS.SESSION,
                     callback: this.removeLoadingState,
@@ -108,8 +110,8 @@ class Expensify extends PureComponent {
     }
 
     render() {
-        // Until the authToken has been initialized from Onyx, display a blank page
-        if (this.state.isLoading) {
+        // Until the authToken has been initialized from Onyx and the onyx migration is done, display a blank page
+        if (this.state.isLoading || !this.state.isOnyxMigrated) {
             return (
                 <View style={styles.genericView} />
             );
